Ensure that we do not repeate names of the members and chinook test update
