+++ src/examples/test_chinook/CMakeLists.txt
@@ -1,0 +1,5 @@
+
+
+add_executable(chinook main.cpp)
+target_link_libraries(chinook PRIVATE Lightweight::Lightweight)
+target_compile_features(chinook PUBLIC cxx_std_23)
