+++ src/examples/test_chinook/entities/Customer.hpp
@@ -1,0 +1,29 @@
+// File is automatically generated using ddl2cpp.
+#pragma once
+
+#include <Lightweight/DataMapper/DataMapper.hpp>
+
+#include "Employee.hpp"
+
+
+struct Customer final
+{
+    static constexpr std::string_view TableName = "Customer";
+
+    Field<int32_t, PrimaryKey::ServerSideAutoIncrement, SqlRealName{"CustomerId"}> CustomerId;
+    Field<SqlUtf16String<40>, SqlRealName{"FirstName"}> FirstName;
+    Field<SqlUtf16String<20>, SqlRealName{"LastName"}> LastName;
+    Field<std::optional<SqlUtf16String<80>>, SqlRealName{"Company"}> Company;
+    Field<std::optional<SqlUtf16String<70>>, SqlRealName{"Address"}> Address;
+    Field<std::optional<SqlUtf16String<40>>, SqlRealName{"City"}> City;
+    Field<std::optional<SqlUtf16String<40>>, SqlRealName{"State"}> State;
+    Field<std::optional<SqlUtf16String<40>>, SqlRealName{"Country"}> Country;
+    Field<std::optional<SqlUtf16String<10>>, SqlRealName{"PostalCode"}> PostalCode;
+    Field<std::optional<SqlUtf16String<24>>, SqlRealName{"Phone"}> Phone;
+    Field<std::optional<SqlUtf16String<24>>, SqlRealName{"Fax"}> Fax;
+    Field<SqlUtf16String<60>, SqlRealName{"Email"}> Email;
+    Field<std::optional<int32_t>, SqlRealName{"SupportRepId"}> SupportRepId;
+
+    BelongsTo<&Employee::EmployeeId, SqlRealName{"SupportRepId"}> c_SupportRepId;
+};
+
