+++ src/examples/test_chinook/entities/Invoiceline.hpp
@@ -1,0 +1,23 @@
+// File is automatically generated using ddl2cpp.
+#pragma once
+
+#include <Lightweight/DataMapper/DataMapper.hpp>
+
+#include "Invoice.hpp"
+#include "Track.hpp"
+
+
+struct Invoiceline final
+{
+    static constexpr std::string_view TableName = "InvoiceLine";
+
+    Field<int32_t, PrimaryKey::ServerSideAutoIncrement, SqlRealName{"InvoiceLineId"}> InvoiceLineId;
+    Field<int32_t, SqlRealName{"InvoiceId"}> InvoiceId;
+    Field<int32_t, SqlRealName{"TrackId"}> TrackId;
+    Field<SqlNumeric<2, 10>, SqlRealName{"UnitPrice"}> UnitPrice;
+    Field<int32_t, SqlRealName{"Quantity"}> Quantity;
+
+    BelongsTo<&Invoice::InvoiceId, SqlRealName{"InvoiceId"}> c_InvoiceId;
+    BelongsTo<&Track::TrackId, SqlRealName{"TrackId"}> c_TrackId;
+};
+
