+++ src/examples/test_chinook/entities/Mediatype.hpp
@@ -1,0 +1,14 @@
+// File is automatically generated using ddl2cpp.
+#pragma once
+
+#include <Lightweight/DataMapper/DataMapper.hpp>
+
+
+struct Mediatype final
+{
+    static constexpr std::string_view TableName = "MediaType";
+
+    Field<int32_t, PrimaryKey::ServerSideAutoIncrement, SqlRealName{"MediaTypeId"}> MediaTypeId;
+    Field<std::optional<SqlUtf16String<120>>, SqlRealName{"Name"}> Name;
+};
+
